+++ mobile/android/android-components/components/concept/push/src/test/java/mozilla/components/concept/push/PushProcessorTest.kt
@@ -49,5 +49,7 @@ class PushProcessorTest {
         override fun onMessageReceived(message: EncryptedPushMessage) {}
 
         override fun onError(error: PushError) {}
+
+        override fun renewRegistration() {}
     }
 }